--- ticketmate-admin/src/main/java/com/ticketmate/backend/admin/core/domain/event/PortfolioHandledEvent.java
@@ -1,8 +1,0 @@
-package com.ticketmate.backend.admin.core.domain.event;
-
-import com.ticketmate.backend.portfolio.core.constant.PortfolioType;
-import java.util.UUID;
-
-public record PortfolioHandledEvent(UUID portfolioId, PortfolioType type) {
-
-}
