--- ticketmate-admin/src/main/java/com/ticketmate/backend/admin/infrastructure/cache/PortfolioCacheEvictListener.java
@@ -1,19 +1,0 @@
-package com.ticketmate.backend.admin.infrastructure.cache;
-
-import com.ticketmate.backend.admin.core.domain.event.PortfolioHandledEvent;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class PortfolioCacheEvictListener {
-
-  /**
-   * PortfolioHandledEvent에서 포트폴리오 승인 시 search 캐시 무효화
-   */
-  @CacheEvict(cacheNames = "searches", allEntries = true)
-  @EventListener
-  public void onPortfolioHandled(PortfolioHandledEvent event) {
-    // AOP가 캐시를 날려주므로 메서드 본문 X
-  }
-}
