+++ ticketmate-admin/src/main/java/com/ticketmate/backend/admin/portfolio/application/dto/request/PortfolioFilteredRequest.java
@@ -1,0 +1,62 @@
+package com.ticketmate.backend.admin.portfolio.application.dto.request;
+
+import com.ticketmate.backend.common.infrastructure.constant.PageableConstants;
+import com.ticketmate.backend.common.infrastructure.util.PageableUtil;
+import com.ticketmate.backend.portfolio.core.constant.PortfolioSortField;
+import com.ticketmate.backend.portfolio.core.constant.PortfolioType;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Direction;
+
+@ToString
+@AllArgsConstructor
+@Getter
+@Setter
+@Builder
+public class PortfolioFilteredRequest {
+
+  private String username; // 사용자 이메일
+
+  private String nickname; // 사용자 닉네임
+
+  private String name; // 사용자 이름
+
+  private PortfolioType portfolioType; // 포트폴리오 타입
+
+  @Min(value = 1, message = "페이지 번호는 1이상 값을 입력해야합니다.")
+  @Max(value = Integer.MAX_VALUE, message = "정수 최대 범위를 넘을 수 없습니다.")
+  private Integer pageNumber; // 페이지 번호 (1부터 시작)
+
+  @Min(value = 1, message = "페이지 당 데이터 최솟값은 1개 입니다.")
+  @Max(value = PageableConstants.MAX_PAGE_SIZE, message = "페이지 당 데이터 최댓값은 " + PageableConstants.MAX_PAGE_SIZE + "개 입니다.")
+  private Integer pageSize; // 페이지 사이즈
+
+  private PortfolioSortField sortField;
+
+  private Sort.Direction sortDirection;
+
+  // 기본값 할당 (1페이지 10개, 최신순)
+  public PortfolioFilteredRequest() {
+    this.pageNumber = 1;
+    this.pageSize = PageableConstants.DEFAULT_PAGE_SIZE;
+    this.sortField = PortfolioSortField.CREATED_DATE;
+    this.sortDirection = Direction.DESC;
+  }
+
+  public Pageable toPageable() {
+    return PageableUtil.createPageable(
+        pageNumber,
+        pageSize,
+        PageableConstants.DEFAULT_PAGE_SIZE,
+        sortField,
+        sortDirection
+    );
+  }
+}
