+++ ticketmate-admin/src/main/java/com/ticketmate/backend/admin/portfolio/infrastructure/event/PortfolioHandledEvent.java
@@ -1,0 +1,8 @@
+package com.ticketmate.backend.admin.portfolio.infrastructure.event;
+
+import com.ticketmate.backend.portfolio.core.constant.PortfolioType;
+import java.util.UUID;
+
+public record PortfolioHandledEvent(UUID portfolioId, PortfolioType type) {
+
+}
