+++ ticketmate-api/src/main/java/com/ticketmate/backend/api/application/controller/admin/AdminControllerDocs.java
@@ -2,15 +2,15 @@ package com.ticketmate.backend.api.application.controller.admin;
 
 import com.chuseok22.apichangelog.annotation.ApiChangeLog;
 import com.chuseok22.apichangelog.annotation.ApiChangeLogs;
-import com.ticketmate.backend.admin.application.dto.request.ConcertHallInfoEditRequest;
-import com.ticketmate.backend.admin.application.dto.request.ConcertHallInfoRequest;
-import com.ticketmate.backend.admin.application.dto.request.ConcertInfoEditRequest;
-import com.ticketmate.backend.admin.application.dto.request.ConcertInfoRequest;
-import com.ticketmate.backend.admin.application.dto.request.PortfolioFilteredRequest;
-import com.ticketmate.backend.admin.application.dto.request.PortfolioStatusUpdateRequest;
-import com.ticketmate.backend.admin.application.dto.response.CoolSmsBalanceResponse;
-import com.ticketmate.backend.admin.application.dto.response.PortfolioFilteredAdminResponse;
-import com.ticketmate.backend.admin.application.dto.response.PortfolioForAdminResponse;
+import com.ticketmate.backend.admin.concerthall.application.dto.request.ConcertHallInfoEditRequest;
+import com.ticketmate.backend.admin.concerthall.application.dto.request.ConcertHallInfoRequest;
+import com.ticketmate.backend.admin.concert.application.dto.request.ConcertInfoEditRequest;
+import com.ticketmate.backend.admin.concert.application.dto.request.ConcertInfoRequest;
+import com.ticketmate.backend.admin.portfolio.application.dto.request.PortfolioFilteredRequest;
+import com.ticketmate.backend.admin.portfolio.application.dto.request.PortfolioStatusUpdateRequest;
+import com.ticketmate.backend.admin.sms.application.dto.response.CoolSmsBalanceResponse;
+import com.ticketmate.backend.admin.portfolio.application.dto.response.PortfolioFilteredAdminResponse;
+import com.ticketmate.backend.admin.portfolio.application.dto.response.PortfolioForAdminResponse;
 import com.ticketmate.backend.auth.infrastructure.oauth2.CustomOAuth2User;
 import com.ticketmate.backend.concert.application.dto.request.ConcertFilteredRequest;
 import com.ticketmate.backend.concert.application.dto.response.ConcertFilteredResponse;
